Demo notebook for multisession registration
This notebook shows how to utilize the function register_multisession(). It explains input arguments and output parameters and shows what to do with the outputs. It specifically shows how to filter registered neurons by their number of sessions where they appear, and how use the output to index traces from the estimates object. Note that because the demo dataset did not include movies or traces, the second part has not been tested specifically with this dataset, but the code should in principal work.

To do:
Maybe add a bit more demos on how to handle the quite big and unintuitive output lists and how to use them for combining/organising the traces.