Collection of short post-CaImAn scripts
These are scripts that people could use after the standard CaImAn pipeline.

The first two functions are for initial results screening:
Check_eval_results() takes indices of components from the estimates object and checks if they got accepted or rejected during evaluation and why. It gives the results in text form as a console printout. The function is just a collection of if-checks, but might be useful for screening the results and to learn how to adapt evaluation parameters.
Plot_component_traces() simply plots individual traces as a line plot as well as a color-coded pcolormesh plot. Useful to get an idea about the general activity of the population, but the number of plotted traces is limited by mainly the figure and screen size, so some pre-selection should be done before.

The other functions are for further analysis, mainly correlation checks (correlation matrix, partial correlation, cross correlation etc.). This can be useful to see whether one neuron was split into two components by accident (these should have very high correlations). These functions are less generalizable than the other two (probably), but I am still leaving them in here, maybe they will be useful sometime. If the post_analysis.py file is in a pythonpath directory, it can be imported and used in other files.

IMPORTANT: The correlation functions use Pandas and Seaborn for calculating correlations and plotting. As far as I know these packages are not included in the standard CaImAn, so would have to be downloaded separately.